Update rest-api-quiz.md
Q24. What should you add to a Cache-Control response header to specify that a response should not be stored in an intermediary cache?
private

None of the other options exist to the Cache-Control response header.

"private
The response may be stored only by a browser's cache, even if the response is normally non-cacheable. If you mean to not store the response in any cache, use no-store instead. This directive is not effective in preventing caches from storing your response."
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control